polymarket-weather-bot/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
py-clob-client>=0.34.0  # necessário apenas para apostas reais (não para DRY_RUN)
